requirements.txt
fastapi
requests
PyGithub
uvicorn[standard]
pydantic
